plain-pages/pyproject.toml
[tool.poetry]
name = "plain-pages"
packages = [
    { include = "plain" },
]

version = "0.0.0"
description = ""
authors = ["Dave Gaeddert <dave.gaeddert@dropseed.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
mistune = "^3.0.0"
python-frontmatter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipdb = "^0.13.9"
isort = "^5.10.1"
black = "^23.1.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
